pyproject.toml
[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[project]
name = "risk-analytics-agent"
version = "0.1.0"
description = "Google ADK multi-agent system for financial liquidity risk analytics"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
license = { text = "MIT" }
keywords = ["risk", "analytics", "agent", "financial", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "google-adk>=1.5.0",
    "litellm>=1.50.0",
    "snowflake-connector-python>=3.12.0",
    "polars>=1.0.0",
    "numpy>=2.0.0",
    "scipy>=1.14.0",
    "matplotlib>=3.9.0",
    "plotly>=5.24.0",
    "kaleido>=0.2.1",
    "jinja2>=3.1.0",
    "weasyprint>=62.0",
    "markdown>=3.7",
    "pyyaml>=6.0",
    "pydantic>=2.9.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/risk-analytics-agent"
Documentation = "https://github.com/yourusername/risk-analytics-agent#readme"
Repository = "https://github.com/yourusername/risk-analytics-agent.git"
Issues = "https://github.com/yourusername/risk-analytics-agent/issues"

[project.scripts]
risk-analytics = "risk_analytics_agent.app:main"

[tool.uv]
package = true

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C", "ARG", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
